+++ webview-ui/src/components/chat/TaskTimeline.tsx
@@ -5,13 +5,13 @@ import { TaskTimelineMessage } from "./TaskTimelineMessage"
 import { MAX_HEIGHT_PX as TASK_TIMELINE_MAX_HEIGHT_PX } from "../../utils/timeline/calculateTaskTimelineSizes"
 import { consolidateMessagesForTimeline } from "../../utils/timeline/consolidateMessagesForTimeline"
 import { calculateTaskTimelineSizes } from "../../utils/timeline/calculateTaskTimelineSizes"
-import { getTaskTimelineMessageColor } from "../../utils/timeline/taskTimelineTypeRegistry"
+import { getTaskTimelineMessageColor } from "../../utils/messageColors"
 import { TooltipProvider } from "../ui/tooltip"
 
 // We hide the scrollbars for the TaskTimeline by wrapping it in a container with
 // overflow hidden. This hides the scrollbars for the actual Virtuoso element
 // by clipping them out view. This just needs to be greater than the webview scrollbar width.
-const SCROLLBAR_WIDTH_PX = 15
+const SCROLLBAR_WIDTH_PX = 25
 
 interface TaskTimelineProps {
 	groupedMessages: (ClineMessage | ClineMessage[])[]